requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Optional (for frontend later)
# streamlit>=1.25.0
# fastapi>=0.100.0
# uvicorn>=0.23.0